requirements.txt
numpy
opencv_python
Pillow
tensorflow
tensorflow_intel
flask
